+++ VyPER/io/__init__.py
@@ -1,0 +1,5 @@
+from .writer import PredictionWriter
+
+__all__ = [
+    'PredictionWriter',
+]