+++ VyPER/utils/__init__.py
@@ -1,8 +1,10 @@
 from .scatter import group_batch
 from .log import get_neutrino_p4, log_hist2D
+from .connectivity import edge_reduction
 
 __all__ = [
     'group_batch',
     'get_neutrino_p4',
     'log_hist2D',
+    'edge_reduction',
 ]